+++ src/VSPackage.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
+
+namespace FileIcons
+{
+    [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
+    [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
+    [Guid(PackageGuids.guidVSPackageString)]
+    [ProvideAutoLoad(UIContextGuids.SolutionHasSingleProject, PackageAutoLoadFlags.BackgroundLoad)]
+    [ProvideAutoLoad(UIContextGuids.SolutionHasMultipleProjects, PackageAutoLoadFlags.BackgroundLoad)]
+    [ProvideMenuResource("Menus.ctmenu", 1)]
+    public sealed class FileIconPackage : Package
+    {
+        protected override void Initialize()
+        {
+            ReportMissingIcon.Initialize(this);
+        }
+    }
+}